Add patched extension release to repo
The quizsliver extension was being detected on CUI Portal. This was due to javascript code present on the quiz page that blocked any extension with the name sliver.js. To counter this detection mechanism, I simply changed the name of sliver.js to extension.js and all occurrences of sliver.js to extension.js in manifest.json and injector.js in the release files of quizsliver. The extension no longer gets detected by the portal now. Happy quizzing.